+++ frontend-web/webclient/app/Utilities/DateUtilities.tsx
@@ -1,6 +1,47 @@
 import {format} from "date-fns/esm";
 
 // Could potentially cause issues with time if user is outside CEST
-export const dateToStringNoTime = (date: number) => format(date, "dd/MM/yyyy");
-export const dateToString = (date: number) => format(date, "HH:mm dd/MM/yyyy");
-export const dateToTimeOfDayString = (date: number) => format(date, "HH:mm");
+export const dateToStringNoTime = (date: number): string => format(date, "dd/MM/yyyy");
+export const dateToString = (date: number): string => format(date, "HH:mm dd/MM/yyyy");
+export const dateToTimeOfDayString = (date: number): string => format(date, "HH:mm");
+
+export const getEndOfDay = (d: Date): Date => {
+    const copy = new Date(d);
+    copy.setHours(23);
+    copy.setMinutes(59)
+    copy.setSeconds(59);
+    copy.setMilliseconds(999);
+    return copy;
+}
+
+export const getStartOfDay = (d: Date): Date => {
+    const copy = new Date(d);
+    copy.setHours(0);
+    copy.setMinutes(0);
+    copy.setSeconds(0);
+    copy.setMilliseconds(0);
+    return copy;
+};
+
+export const getStartOfWeek = (d: Date): Date => {
+    const day = d.getDay();
+    const diff = d.getDate() - day + (day === 0 ? -6 : 1);
+
+    const copy = new Date(d);
+    copy.setDate(diff);
+    copy.setHours(0);
+    copy.setMinutes(0);
+    copy.setSeconds(0);
+    copy.setMilliseconds(0);
+    return copy;
+};
+
+export const getStartOfMonth = (d: Date): Date => {
+    const copy = new Date(d);
+    copy.setDate(1);
+    copy.setHours(0);
+    copy.setMinutes(0);
+    copy.setSeconds(0);
+    copy.setMilliseconds(0);
+    return copy;
+}
